Fix some bugs Improve code Update README Update project infos

+++ resources/ts/components/TodoItem.tsx
@@ -1,5 +1,5 @@
 import { React, useContext, useRef, useState } from "react";
-import { ListGroup, Form, Stack, Button } from "react-bootstrap";
+import { ListGroup, Form, Stack, Button, Spinner } from "react-bootstrap";
 import { TodoType } from "../interfaces/ index";
 import { TodoContext } from "../contexts/TodoContext";
 
@@ -12,53 +12,106 @@ export default function TodoItem({ todo }: Props) {
 
     const [checked, setChecked] = useState<boolean>(todo.completed as boolean);
     const [contentEditable, setContentEditable] = useState<boolean>(false);
-    const descriptionRef = useRef<HTMLLabelElement | null>(null);
+    const labelRef = useRef<HTMLLabelElement | null>(null);
+    const [updating, setUpdating] = useState<boolean>(false)
+    const [deleting, setDeleting] = useState<boolean>(false)
 
-    const handleSetChecked = () => {
-        setChecked(!checked);
+    const handleSetChecked = async (): Promise<void> => {
+        setUpdating(true)
+
+        const res = await fetch("http://127.0.0.1:8000/api/todos/" + todo.id, {
+            method: "put",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ completed: !checked }),
+        })
+
+        if (res.status !== 200) {
+            setUpdating(false)
+            return alert("Fail to update")
+        }
 
         dispatch({
             type: "update",
             payload: {
                 id: todo.id,
-                completed: checked,
+                completed: !checked,
             },
         });
+
+        setUpdating(false)
+        setChecked(!checked);
     };
 
-    const handleSetContentEditable = () => {
-        setContentEditable(!contentEditable);
+    const handleSetContentEditable = async (): Promise<void> => {
+        if (!contentEditable) {
+            setContentEditable(true)
+        }
 
         if (contentEditable) {
+            if (labelRef.current!.innerText === "") return
+
+            setUpdating(true)
+
+            const res = await fetch("http://127.0.0.1:8000/api/todos/" + todo.id, {
+                method: "put",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify({ description: labelRef.current!.innerText }),
+            })
+
+            if (res.status !== 200) {
+                setUpdating(false)
+                return alert("Fail to update")
+            }
+
             dispatch({
                 type: "edit",
                 payload: {
                     id: todo.id,
-                    description: descriptionRef.current!.innerText,
+                    description: labelRef.current!.innerText,
                 },
             });
+
+            setUpdating(false)
+            setContentEditable(false)
         }
     };
 
-    const handleDelete = () => {
+    const handleDelete = async (): Promise<void> => {
+        if (!confirm("Are you sure you want to delete this todo?")) return
+
+        setDeleting(true)
+
+        const res = await fetch("http://127.0.0.1:8000/api/todos/" + todo.id, {
+            method: "delete",
+            headers: { "Content-Type": "application/json" },
+        })
+
+        if (res.status !== 200) {
+            setDeleting(false)
+            return alert("Fail to update")
+        }
+
         dispatch({
             type: "delete",
             payload: {
                 id: todo.id,
             },
         });
+
+        setDeleting(false)
     };
 
     return (
         <ListGroup.Item className="d-flex align-items-center justify-content-between">
             <Form.Check>
-                <Form.Check.Input
+                {!updating && <Form.Check.Input
                     checked={checked}
                     onChange={handleSetChecked}
                     type="checkbox"
-                />
+                />}
+                {updating && <Spinner variant="primary" animation="border" size="sm" className="me-3" />}
                 <Form.Check.Label
-                    ref={descriptionRef}
+                    ref={labelRef}
                     className={checked ? "text-decoration-line-through" : ""}
                     contentEditable={contentEditable}
                 >
@@ -76,10 +129,11 @@ export default function TodoItem({ todo }: Props) {
                     onClick={handleSetContentEditable}
                 >
                     {!contentEditable && "Edit"}
-                    {contentEditable && "Save"}
+                    {contentEditable ? (!updating ? "Save" : <Spinner animation="border" size="sm" />) : null}
                 </Button>}
                 <Button variant="outline-danger" size="sm" onClick={handleDelete}>
-                    Delete
+                    {!deleting && "Delete"}
+                    {deleting && <Spinner animation="border" size="sm" />}
                 </Button>
             </Stack>
         </ListGroup.Item>
